requirements/base.txt
pytz==2019.2  # https://github.com/stub42/pytz
python-slugify==3.0.4  # https://github.com/un33k/python-slugify
Pillow==6.1.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.4  # https://github.com/evansd/whitenoise
redis==3.3.8  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==2.2.5  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.40.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.3  # https://github.com/django-compressor/django-compressor
django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-log-request-id==1.3.2
django-nested-admin==3.2.4
pytest-django==3.5.1
django-allauth==0.40.0
django-prometheus==1.0.15
django-cors-middleware==1.4.0
django-axes==5.0.13
django-redis-cache==2.0.
django-bootstrap-form==3.4
django-extensions==2.2.1  # https://github.com/django-extensions/django-extensions


# Django REST Framework
djangorestframework==3.10.3  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

-e git://github.com/enkore/django-rest-auth.git@UnignoreMoFiles#egg=django-rest-auth
hiredis==1.0.0

pprintpp==0.4.0
-e git+ssh://git@git.qabel.de/Qabel/qabel-web-theme.git@0.1.6#egg=qabel-web-theme

pytz==2019.2
